# Refine along fitted line for faster BC1 encoding