--- a/src/encode/bc.rs
+++ b/src/encode/bc.rs
@@ -153,10 +153,12 @@ fn get_bc1_options(options: &EncodeOptions) -> bc1::Bc1Options {
         dither: options.dithering.color(),
         perceptual: options.error_metric == ErrorMetric::Perceptual,
         opaque_always_p4: options.quality <= CompressionQuality::Normal,
+        refine: options.quality >= CompressionQuality::Normal,
         refine_max_iter: match options.quality {
             CompressionQuality::Fast => 0,
-            CompressionQuality::Normal => 3,
-            CompressionQuality::High | CompressionQuality::Unreasonable => 10,
+            CompressionQuality::Normal => 1,
+            CompressionQuality::High => 5,
+            CompressionQuality::Unreasonable => 10,
         },
         quantization: match options.quality {
             CompressionQuality::Fast => bc1::Quantization::ChannelWiseOptimized,