… test-data/encode_quality.md
… test-data/output-encode/base/BC1_UNORM.dds
… test-data/output-encode/base/BC2_UNORM.dds
… test-data/output-encode/base/BC2_UNORM_PREMULTIPLIED_ALPHA.dds
… test-data/output-encode/base/BC3_UNORM.dds
… test-data/output-encode/base/BC3_UNORM_NORMAL.dds
… test-data/output-encode/base/BC3_UNORM_PREMULTIPLIED_ALPHA.dds
… test-data/output-encode/base/BC3_UNORM_RXGB.dds
… test-data/output-encode/base/_hashes.yml
… test-data/output-encode/dither/BC1_UNORM base.dds
… test-data/output-encode/dither/BC2_UNORM base.dds
… test-data/output-encode/dither/BC2_UNORM_PREMULTIPLIED_ALPHA base.dds
… test-data/output-encode/dither/BC3_UNORM base.dds
… test-data/output-encode/dither/BC3_UNORM_NORMAL base.dds
… test-data/output-encode/dither/BC3_UNORM_PREMULTIPLIED_ALPHA base.dds
… test-data/output-encode/dither/BC3_UNORM_RXGB base.dds
… test-data/output-encode/dither/_hashes.yml