This PR implements an idea I had a while ago to make BC1 encoding faster.

The reason BC1 encoding is so slow is that the refinement algorithm I implemented is very expensive for 3 color channels (dimensions). However, it's quite fast for 1 dimension. So the idea is: instead of trying out endpoint variations in 3D space, vary the distance of the endpoints from their midpoint.

This works surprisingly well. It can't reach the quality of refinement in 3D space, but it also doesn't have to. Since line refinement is about 3x faster than 3D refinement, I can use line refinement to cut down the number of iterations 3D color refinement has to do. So I can improve performance, while keeping the quality the same.

And that's exactly what I did. I changed the parameters of BC1 encoding to have the same level of quality as before, while being faster thanks to line refinement:

| Quality           | Old       | New       | Change           |
| ----------------- | --------- | --------- | ---------------- |
| Fast              | 990.94 µs | 986.58 µs | -1.3958% (noise) |
| Normal            | 5.5859 ms | 4.0678 ms | -27.176%         |
| High              | 15.690 ms | 14.363 ms | -8.4546%         |
| Perceptual Normal | 12.392 ms | 8.6790 ms | -29.964%         |

(Fast quality isn't affected, because it doesn't do any refinement.)