@RunDevelopment 2025-10-07
I spend some more time integrating the least square optimizing from #79 into this PR. This worked quite well. The encoder is now faster and produces images of higher quality. A win on all fronts.

| Quality           | Old       | New       | Change           |
| ----------------- | --------- | --------- | ---------------- |
| Fast              | 726.89 µs | 731.83 µs | +0.8483% (noise) |
| Normal            | 4.2644 ms | 2.1769 ms | -48.952%         |
| High              | 11.936 ms | 11.433 ms | -4.2141%         |
| Perceptual Normal | 7.3608 ms | 3.5639 ms | -51.583%         |

The only "downside" is that the speed improvement for High quality isn't as much. But I think that's a reasonable tradeoff for better quality.